Rename rhel90 to rhel9
The rhel9 package defines RHEL 9.0, RHEL 9.1, and will define all minor
RHEL 9.y versions.  The package name now reflects this.
