+++ internal/distro/rhel9/distro.go
@@ -1,4 +1,4 @@
-package rhel90
+package rhel9
 
 import (
 	"encoding/json"
